chore(frontend): move size-screen.contant.ts to utils
